replace "tty" command w. "test -t 0", MacOS X color fix
I'm optimistic that most UNIX and UNIX-like systems support "test -t 0",
but the possibility exists that some systems will not like this.

This fix allows color to work in MacOS X 10.5.7. For some reason, the
tty command is flaky when run from ~/.bash_profile and will report no
tty. Didn't quite figure it out. But I did isolate it to this
conditional, and verified that "test -t 0" (which is recommended as
an alternative in the MacOS X "tty" man page) does get MacOS X color
working again.