FL-2195: From Todd Eigenschink <todd@xymmetrix.com>:
OpenSSH 6.8, which was just released, includes the following change:

http://www.openssh.com/txt/release-6.8

   Fingerprints now have the hash algorithm prepended. An example of
   the new format: SHA256:mVPwvezndPv/ARoIadVY98vAC0g+P/5633yTC4d/wXE
   Please note that visual host keys will also be different.

(Previous versions always dumped in MD5 with no prefix.)

The pattern match that keychain uses doesn't work with the new
fingerprint format. The patch below takes care of it.

I used shopt extglob so I could use @( ) for the alternative. I'm not
a bash pattern whiz; there may be a better way to do that.