Add ability to load all keyfiles in `ssh_config`
When using the `--confhost` option one can only specify a single host to
load a key from. However one also has to specify for which host to load
the keys from. When you want to load several keyfiles that belong to
several hosts into the keychain, this requires many functioncalls, and
is cumbersome.

I think it should be able to load all the hosts specified in the
`ssh_config` with a single command. Normally the config contains
information about all the hosts that you have keyfiles for, so it would
make sence that a person wants to add all these keys to their keychain
for ease of use.

With some minor adjustments to the way `pkeypath` is being used in the
script, its now able to load all the hosts from the configuration file
with a single function call. Personally I will use this during boot to
load all my ssh-keys into my agent so that upon entering my password,
all my servers will be accesible.