Refuse to process malformed SSH public key files.
If the user specifies a file that looks like an SSH public key file,
but it does not contain exactly one SSH public key (either it contains
none or the fingerprint is wrong or there is another key appended at
the end), keychain will misinterpret the "fingerprint" of the key file
and always try to load the corresponding private key.  This may lead
to repeated passphrase prompts and general user confusion.

A trivial way to reproduce this is to append a second line containing
the fingerprint of another SSH public key to a keyfile.

Debian bug:	https://bugs.debian.org/673019
Reported by:	Ryan Kavanagh <rak@debian.org>