Let's rework some things:
Deprecating the --inherit option, as it doesn't make sense for
gpg-agent and is a confusing option. Instead, maintain the same
default behavior for keychain, and add a --ssh-allow-forwarded
to allow forwarded connections if needed. Preserve the --noinherit
option.

Also, make it clear in the docs that --stop only applies to
ssh-agent and remove support for gpg-agent. Combine this with
a rewrite of stopagent() (now called stop_ssh_agents()) which
is a lot easier to read.