Testing some interesting new approaches.
I dropped the --agents option, which is now deprecated. The code
now assumes you want an ssh-agent, even if you don't specify keys.
But it assumes you don't want gpg-agent, unless you specify gpg
keys on the command-line.

The --clear option can only be used in conjunction with adding
keys, not standalone.

By default, official ssh-agent is used, unless you specify the
--ssh-use-gpg option, which will allow gpg-agent to be used. So
it's an opt-in.

This makes the code even simpler, at 1127 lines, and also seems
more intuitive to me.